requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
Django==4.1.4
django-rest-knox==4.2.0
djangorestframework==3.14.0
kombu==5.2.4
prompt-toolkit==3.0.36
pycparser==2.21
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
timedelta==2020.12.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
